+++ apps/teka/src/routes/changelog/+page.svelte
@@ -53,7 +53,11 @@
 <h3>Correcciones</h3>
 <ul>
     <li>Corregido algún error menor con el buscador.</li>
-    <li>Resulta que las analíticas de Vercel no estaban funcionando.</li>
+    <li>Resulta que las analíticas de Vercel no estaban funcionando, ahora sí lo hacen.</li>
+    <li>
+        Ahora el buscador permite usar filtros sin escribir para mostrar todos los videos (como
+        debería, cosa que no hacía).
+    </li>
     <li>
         De hecho, el naranja que usaba la web no era exactamente el mismo que el del logo, esto se
         ha corregido.