+++ CHANGELOG.md
@@ -21,7 +21,8 @@
 ### Correcciones
 
 - Corregido algún error menor con el buscador.
-- Resulta que las analíticas de Vercel no estaban funcionando.
+- Resulta que las analíticas de Vercel no estaban funcionando, ahora sí lo hacen.
+- Ahora el buscador permite usar filtros sin escribir para mostrar todos los videos (como debería, cosa que no hacía).
 - [ZPW/TEKA] De hecho, el naranja que usaba la web no era exactamente el mismo que el del logo, esto se ha corregido.
 - [ZPW/PROFE] [Apps/Calculadora de Nóminas] Corregido un fallo en que el primer grupo de cotización no era usable.